Separate prototype routes
